Revert "capsicum: print seconds immediately"
This reverts commit 6404c338641d8b45049232ae36df0391cbceaa68.

Minimize the capsicum sandbox: disallow fstat()/ioctl() again.

The fstat()/ioctl() calls check if the process is connected to a tty,
and if so, enables line buffering. Since these calls were disallowed:

* tty test failed
* process is not attached to a tty
* output is fully buffered

In the case where pseudocron prints out the remaining time and sleeps,
when the output occurs doesn't matter because the calling process is
blocked anyway.